Move Repo and Com to src
